Version 1.0: piMusicPlayer running on Linux (Debian).
It is a subset of piMusicSystem. This music player on a Raspberry-pi can be controlled locally with buttons mounted on LCD module. The LCD module is mounted on top of the Raspberry-pi. The buttons and LCD module are controlled by separate project. You can control the player locally with a bash script. With an Android application which is a separate project, you can control the player remotely.